Update dag-json codec.
