+++ dag-json/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Irakli Gozalishvili <contact@gozala.io>"]
 edition = "2018"
 
 [dependencies]
-base64 = "0.12.0"
+base64 = "0.12.1"
 libipld-core = { path = "../core" }
-serde_json = "1.0.50"
+serde_json = "1.0.52"
 serde = { version = "1.0.106", features = ["derive"] }
